Add services to support statsController.js
1. Add a Team service to update $scope.teams. This array variable is use
to fill the Team Ranking Section.
- create a teamService.js file containing a factory with the "teams"
array of objects.
- This team array of objects has a method to return this object.
2. Add a CurrentWeek  service to update the scope currentWeek variable.
- create a currentWeekService.js file containing a service to return the
current week of the challenge.
- The getCurrentWeek function takes the start date of the challenge
minus today's Date. This amount is placed in numberOfDays variable then
divide by 7 to get currentWeek.
- The service provides a method to return this number.
3. Add a Players service to update the scope currentPlayer object
variable. This object's properties is use to fill the jerseyNumber and
teamName in the Player Profile section, then the playerTotalPoints
variable in the Player stats section.
- create a playersService.js file containing a factory with the
"players" array of objects and currentPlayer object.
- This team array of objects has a method to return the players object,
return the currentPlayer object, and update the currentPlayer object.
- the method to update the currentPlayer object use a for-loop to match
a jerseyNumber to an object in the players array. When found, that
object is use to update the currentPlayer object.