Add getPlayerTeamRank to Players Service
1. Create a method in the Players service that gets all the team members
of the current player from the players array to find the current rank of
the current player compared to other team members.
- use a for loop in the players array to place all players objects
matching the teamName of the currentPlayer object into a playersOnTeam
array of objects.
- decending sort the playersOnTeam array using a array.sort compare
method.
- use a for loop to search the sorted playersOnTeam array for the
jerseyNumber of the currentPlayer object. If found, use the
array.indexOf to get the current index plus 1 to get the rank.