LogIn service, page, controller and unit test
1. Create a logIn service that checks against a database of log-In
objects, returns jerseyNumber to playerService, and returns a true/false
to open app. If true, the player is sent to the inpute page. If false, a
pass variable is sent to the logIn page to release a failed login
attempt message.
- create a credentials database array to hold the jerseyNumbers and
passwords of current players.
- create a pass variable that will be sent to the logIn controller.
- logIn function that takes parameters from the logIn page and validated
them to a element in the credential array.
- If a match is found, use a Players service method to update the global
jeresyNumber variable in the service. The state is  change to the input
page.
- If no match, a pass variable is return with false. The logIn page will
display a failed attempt.
- a getPass function is create to send that failed attempt message.
2. The logIn page was updated to reflect the jerseyNumber and password
parameters. The logIn function was change to signIn().
3. A logIn controller with the LogIn service was created to provide the
signIn function and setup the failed messsge attempted.
- A signIn function was made to wrap the LogIn service logIn method.
This method returns a false if no match to the  credential array is
found. If a false is return, then the message is sent.
4. A html page with the jasmine unit testing files and a copy of the
logIn service was made.
- Unit test was made to validate only a correct jerseyNumber and
password combination would return a true statement.
5. Player services and its unit testing spec sheet was updated.
- jerseyNumber can now be updated and returned from the Players
services.
- The unit testing was change to reflected this.
5. The app. js file was amended to give the logIn state a controller.